refactor(ngm): move ngm to src root, clean package.json scripts, fix e2e test
