+++ src/@types/cpy/package.json
