+++ src/commands/test.ts
