+++ src/commands/version.command.ts
