+++ src/helpers/inline-resources.ts
