+++ src/lib/ngm.ts
