+++ src/tasks/bundle-umd.task.ts
