+++ src/tasks/npm/build-pkg-json.task.ts
