+++ src/tasks/prepublish-git-check.task.ts
