+++ src/utils/merge-package-json.ts
