+++ src/utils/submodules-resolution.ts
